+++ src/components/common/toggle-group/g-toggle-button-group.scss
@@ -13,7 +13,7 @@
     }
     &.MuiToggleButton-root{
         padding: 5px 10px;
-        margin-right: 1rem;
+        margin-right: .5rem;
         &.Mui-selected{
             color: #ffffff,;
             background-color: #1875d1;
