+++ src/container/settings/Settings.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom'
 import NotificationView from '../../components/notification-view/NotificationView';
-import GroundUpDrawer from '../../components/common/drawer/Drawer'
+import GroundUpDrawer from '../../components/common/drawer/GDrawer'
 import GroundUpAppBar from '../../components/common/header/GroundUpAppBar'
 import { USER_ADMIN, USER_ADMIN_ACTIVITY_MANAGEMENT, USER_ADMIN_CRANE_BOOKING_MANAGEMENT, USER_ADMIN_MAINTENANCE_MANAGEMENT, USER_ADMIN_MATERIAL_MANAGEMENT, USER_ADMIN_REPORTS_MANAGEMENT, USER_ADMIN_ROLE_MANAGEMENT, USER_ADMIN_USER_MANAGEMENT, USER_NOTIFICATION } from '../../constants/ContextPaths';
 import AdminView from '../../components/admin-view/AdminView';
